seo - add google site verification marker file
It looks a little ugly in the src/ directory along with the other files
being nicely named. This is probably better than the alternative of
adding a meta tag to every page by overriding theme/head.hbs. The other
alternatives for verification are DNS records and google analytics, both
of which are not feasible or desirable at the moment.